+++ ta4j-core/src/main/java/org/ta4j/core/indicators/RSIEnergyIndicator.java
@@ -1,0 +1,79 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017-2025 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.indicators;
+
+import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.averages.MMAIndicator;
+import org.ta4j.core.indicators.helpers.GainIndicator;
+import org.ta4j.core.indicators.helpers.LossIndicator;
+import org.ta4j.core.num.Num;
+
+import static org.ta4j.core.num.NaN.NaN;
+
+/**
+ * Relative strength index energy indicator.
+ *
+ * 7
+ */
+public class RSIEnergyIndicator extends CachedIndicator<Num> {
+
+    private final MMAIndicator averageGainIndicator;
+    private final MMAIndicator averageLossIndicator;
+    private final int barCount;
+
+    /**
+     * Constructor.
+     *
+     * @param indicator the {@link Indicator}
+     * @param barCount  the time frame
+     */
+    public RSIEnergyIndicator(Indicator<Num> indicator, int barCount) {
+        super(indicator);
+        this.averageGainIndicator = new MMAIndicator(new GainIndicator(indicator), barCount);
+        this.averageLossIndicator = new MMAIndicator(new LossIndicator(indicator), barCount);
+        this.barCount = barCount;
+    }
+
+    @Override
+    protected Num calculate(int index) {
+        if (index < getCountOfUnstableBars()) {
+            return NaN;
+        }
+        // compute relative strength
+        Num averageGain = averageGainIndicator.getValue(index);
+        Num averageLoss = averageLossIndicator.getValue(index);
+        final var numFactory = getBarSeries().numFactory();
+        if (averageLoss.isZero()) {
+            return averageGain.isZero() ? numFactory.zero() : numFactory.hundred();
+        }
+        Num relativeStrength = averageGain.dividedBy(averageLoss);
+        // compute relative strength index
+        return numFactory.hundred().minus(numFactory.hundred().dividedBy(numFactory.one().plus(relativeStrength)));
+    }
+
+    @Override
+    public int getCountOfUnstableBars() {
+        return this.barCount;
+    }
+}