+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 
 ### Added
 - Bars can now be built by `beginTime` instead of `endTime`
+- Added new 'RSIEnergyIndicator' indicator class
+- Added `substitute` static helper to `UnaryOperator` to replace a specific indicator value with another
 
 
 ## 0.18 (released May 15, 2025)